Saving 2021 updates for universal binary. Not all completed yet.

+++ VSVisibleItemArrayController.m
@@ -1,7 +1,7 @@
 /*********************************************************************************                                                                       
  * The MIT License (MIT)
  * 
- * Copyright (c) 2009-2016 Jason Neuswanger
+ * Copyright (c) 2009-2021 Jason Neuswanger
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -37,7 +37,7 @@
 
 - (void)remove:(id)sender
 {
-	int newSelectionIndex;
+	long newSelectionIndex;
 	if ([self canSelectNext]) {							// If there's another object below the one being deleted, prepare to select it
 		newSelectionIndex = [self selectionIndex];
 	} else {											// If this was the bottom object, then select the next one up
